move read_manifests_if_cached to a new file recipes/utils.py to circumvent circular imports
